fix: update scorer within run
