+++ packages/core/src/executors/index.ts
@@ -1,0 +1,1 @@
+export * from "./run";
