+++ packages/core/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./run";
+export * from "./executors";
 export * from "./stats";
